Encode read-only filesystem for /usr
We take our sysroot, and copy the contents and metadata into
a filesystem directory, where filesystem/meta and filesystem/data
have files meta/123 and data/123 for each file, directory, or
symlink in the copied directory, where 123 is the inode of the
concerned file.

This gives a completely statically-hosted representation of a
read-only filesystem.

(Except for mime-types, those could still give problems.
The server needs to allow the client to request some specific
mime-types, such as application/wasm.)